pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yaloader"
version = "0.0.3"
description = "A package loading class configurations from yaml files"
readme = "README.md"
license = "AGPL-3.0-only"
requires-python = ">=3.9"
authors = [
    { name = "Dominik Mairhöfer", email = "dominik@mairhoefer.xyz" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
]
dependencies = [
    "pydantic>=2.0",
    "pyyaml",
]

[project.urls]
Homepage = "https://github.com/DominikMa/yaloader"
Documentation = "https://yaloader.readthedocs.io"
"Bug Tracker" = "https://github.com/DominikMa/yaloader/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py39"
line-length = 120
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "ANN", "B", "UP", "SIM", "TCH", "RUF"]
ignore = ["ANN401"]  # Any is intentionally used in dynamic dispatch methods

[tool.ruff.lint.per-file-ignores]
"src/yaloader/__init__.py" = ["I001"]  # import order is intentional (circular imports)
"tests/**" = ["ANN"]  # test functions don't need type annotations

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = ["long: marks tests as long-running"]

[tool.coverage.run]
source = ["yaloader"]

[tool.coverage.report]
fail_under = 80

[dependency-groups]
dev = [
    "autodoc-pydantic>=2.2.0",
    "coverage>=7.10.7",
    "furo>=2025.12.19",
    "jupyter-server>=2.17.0",
    "myst-nb>=1.3.0",
    "pytest>=8.4.2",
    "ruff>=0.15.5",
    "sphinx>=7.4.7",
    "sphinx-remove-toctrees>=1.0.0.post1",
]
